+++ src/locale/locales/es/common.json
@@ -199,7 +199,7 @@
     "key_added": "Clave de acceso añadida",
     "key_deleted": "Clave de acceso borrada",
     "unknown_action": "Acción desconocida",
-    "recent_activity": "Últimos movimientos",
+    "recent_activity": "Actividad reciente",
     "collection": "Colección",
     "choose_what_to_send": "Elige que quieres enviar",
     "select_asset": "Elige entre tus activos",
