+++ src/module/transaction/component/feedback/SendTransactionModal/SendTransactionModal.tsx
@@ -1,22 +1,9 @@
 import LoadingModal from "module/common/component/feedback/LoadingModal/LoadingModal";
-import { BaseUseModalWrapperReturn } from "module/common/hook/useModalWrapper";
 import { useTranslate } from "module/common/hook/useTranslate";
 import ConfirmPinModal from "module/settings/components/core/ConfirmPinModal/ConfirmPinModal";
 import { useState } from "react";
-import { UseMutationResult } from "react-query";
-
-export type SendTransactionModalChildrenProps<TData = unknown, TError = unknown, TVariables = void, TContext = unknown> = Pick<
-    BaseUseModalWrapperReturn,
-    "showModal"
-> &
-    Pick<UseMutationResult<TData, TError, TVariables, TContext>, "isLoading" | "isError" | "isSuccess">;
-
-export interface SendTransactionModalProps<TData = unknown, TError = unknown, TVariables = void, TContext = unknown> {
-    useMutationResult: Omit<UseMutationResult<TData, TError, TVariables, TContext>, "mutateAsync">;
-    onExited?: () => unknown;
-    sendTransaction: () => void | Promise<unknown>;
-    children: (props: SendTransactionModalChildrenProps) => JSX.Element;
-}
+import { InteractionManager } from "react-native";
+import { SendTransactionModalProps } from "./SendTransactionModal.types";
 
 function SendTransactionModal<TData = unknown, TError = unknown, TVariables = void, TContext = unknown>({
     useMutationResult,
@@ -28,15 +15,27 @@ function SendTransactionModal<TData = unknown, TError = unknown, TVariables = vo
     const [showConfirmation, setShowConfirmation] = useState(false);
     const { isLoading, isSuccess, isError } = useMutationResult;
 
+    function runAfterInteractions(cb?: () => void) {
+        if (cb) {
+            InteractionManager.runAfterInteractions(() => {
+                cb();
+            });
+        }
+    }
+
+    const handleOnPinConfirmed = () => runAfterInteractions(sendTransaction);
+
+    const handleOnExited = () => runAfterInteractions(onExited);
+
     return (
         <>
             {children({ showModal: () => setShowConfirmation(true), isError, isSuccess, isLoading: isLoading || showConfirmation })}
-            <ConfirmPinModal open={showConfirmation} onClose={() => setShowConfirmation(false)} onPinConfirmed={sendTransaction} />
+            <ConfirmPinModal open={showConfirmation} onClose={() => setShowConfirmation(false)} onPinConfirmed={handleOnPinConfirmed} />
             <LoadingModal
                 loading={isLoading}
                 success={isSuccess}
                 error={isError}
-                onExited={onExited}
+                onExited={handleOnExited}
                 successMessage={translate("transaction_completed")}
             />
         </>
